move top level page layouts for each route from "routes" to "pages"
Free up the routes location for use by react-router in framework mode
